+++ packages/authhero/test/routes/management-api/forms.spec.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from "vitest";
 import { testClient } from "hono/testing";
 import { getAdminToken } from "../../helpers/token";
 import { getTestServer } from "../../helpers/test-server";
-import { Form, FormFieldType, FormType } from "@authhero/adapter-interfaces";
+import { Form } from "@authhero/adapter-interfaces";
 
 describe("forms", () => {
   it("should support crud", async () => {
@@ -17,19 +17,6 @@ describe("forms", () => {
       {
         json: {
           name: "signup",
-          type: FormType.SIGNUP,
-          fields: [
-            {
-              type: FormFieldType.TEXT,
-              name: "email",
-              id: "email",
-              label: "Email",
-              required: true,
-              readOnly: false,
-              disabled: false,
-              visible: true,
-            },
-          ],
         },
         header: {
           "tenant-id": "tenantId",
@@ -46,13 +33,6 @@ describe("forms", () => {
     const { created_at, updated_at, id, ...rest } = createdForm;
     expect(rest).toMatchObject({
       name: "signup",
-      type: "signup",
-      fields: [
-        expect.objectContaining({
-          name: "email",
-          label: "Email",
-        }),
-      ],
     });
     expect(created_at).toBeTypeOf("string");
     expect(updated_at).toBeTypeOf("string");