+++ packages/kysely/test/forms/create.spec.ts
@@ -1,10 +1,5 @@
 import { describe, expect, it } from "vitest";
 import { getTestServer } from "../helpers/test-server";
-import {
-  FormFieldType,
-  FormInsert,
-  FormType,
-} from "@authhero/adapter-interfaces";
 
 describe("forms", () => {
   describe("create", () => {
@@ -21,29 +16,11 @@ describe("forms", () => {
 
       const createdForm = await data.forms.create("tenantId", {
         name: "Basic Form",
-        type: FormType.CUSTOM,
-        fields: [
-          {
-            id: "field1",
-            type: FormFieldType.TEXT,
-            label: "Name",
-            required: true,
-            name: "field1",
-          },
-        ],
       });
 
       expect(createdForm).toMatchObject({
         id: expect.any(String),
         name: "Basic Form",
-        fields: [
-          {
-            id: "field1",
-            type: FormFieldType.TEXT,
-            label: "Name",
-            required: true,
-          },
-        ],
       });
 
       // Verify the form was actually created by retrieving it
@@ -51,14 +28,6 @@ describe("forms", () => {
       expect(retrievedForm).toMatchObject({
         id: createdForm.id,
         name: "Basic Form",
-        fields: [
-          {
-            id: "field1",
-            type: "text",
-            label: "Name",
-            required: true,
-          },
-        ],
       });
     });
 
@@ -73,52 +42,8 @@ describe("forms", () => {
         sender_name: "SenderName",
       });
 
-      const formData: FormInsert = {
+      const formData = {
         name: "Complex Form",
-        type: FormType.CUSTOM,
-        fields: [
-          {
-            id: "field1",
-            type: FormFieldType.TEXT,
-            label: "Full Name",
-            placeholder: "Enter your full name",
-            required: true,
-            name: "field1",
-          },
-          {
-            id: "field2",
-            type: FormFieldType.EMAIL,
-            label: "Email Address",
-            placeholder: "your.email@example.com",
-            required: true,
-            name: "field2",
-            // validations: {
-            // pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
-            // },
-          },
-          {
-            id: "field3",
-            type: FormFieldType.SELECT,
-            label: "Country",
-            options: [
-              { value: "us", label: "United States" },
-              { value: "ca", label: "Canada" },
-              { value: "uk", label: "United Kingdom" },
-            ],
-            required: false,
-            name: "field3",
-          },
-        ],
-        controls: [
-          {
-            id: "controlsId",
-            label: "Form Controls",
-            type: "submit",
-          },
-        ],
-        layout: {
-          columns: 1,
-        },
       };
 
       const createdForm = await data.forms.create("tenantId", formData);