+++ packages/kysely/test/forms/list.spec.ts
@@ -1,7 +1,5 @@
 import { describe, expect, it } from "vitest";
 import { getTestServer } from "../helpers/test-server";
-import { FormType } from "@authhero/adapter-interfaces";
-import { FormFieldType } from "@authhero/adapter-interfaces";
 
 describe("forms", () => {
   describe("list", () => {
@@ -18,38 +16,16 @@ describe("forms", () => {
 
       const form = await data.forms.create("tenantId", {
         name: "Basic Form",
-        type: FormType.CUSTOM,
-        fields: [
-          {
-            id: "field1",
-            type: FormFieldType.TEXT,
-            label: "Name",
-            required: true,
-            name: "field1",
-          },
-        ],
       });
 
       const result = await data.forms.list("tenantId");
 
       expect(result).toMatchObject({
         forms: [
-          {
+          expect.objectContaining({
             id: form.id,
             name: "Basic Form",
-            fields: [
-              {
-                disabled: false,
-                id: "field1",
-                type: "text",
-                label: "Name",
-                required: true,
-                name: "field1",
-                readOnly: false,
-                visible: true,
-              },
-            ],
-          },
+          }),
         ],
         length: 1,
         limit: 50,
@@ -69,30 +45,10 @@ describe("forms", () => {
 
       const form1 = await data.forms.create("tenantId", {
         name: "Registration Form",
-        type: FormType.CUSTOM,
-        fields: [
-          {
-            id: "field1",
-            name: "name",
-            type: FormFieldType.TEXT,
-            label: "Name",
-            required: true,
-          },
-        ],
       });
 
       const form2 = await data.forms.create("tenantId", {
-        type: FormType.CUSTOM,
         name: "Contact Form",
-        fields: [
-          {
-            id: "field1",
-            name: "name",
-            type: FormFieldType.TEXT,
-            label: "Message",
-            required: true,
-          },
-        ],
       });
 
       const result = await data.forms.list("tenantId");